move NetworkMessageSerializer
