+++ src/Network/NetworkMessage.php
@@ -5,7 +5,7 @@ namespace BitWasp\Bitcoin\Network;
 use BitWasp\Buffertools\Buffer;
 use BitWasp\Bitcoin\Crypto\Hash;
 use BitWasp\Bitcoin\Serializable;
-use BitWasp\Bitcoin\Serializer\NetworkMessageSerializer;
+use BitWasp\Bitcoin\Serializer\Network\NetworkMessageSerializer;
 
 class NetworkMessage extends Serializable
 {
